collapse Producers
